pyproject.toml
[tool.poetry]
name = "migrate"
version = "0.1.0"
description = ""
authors = ["theKwabena <abbeyemmanueltetteh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
requests = "^2.31.0"
celery = {extras = ["redis"], version = "^5.3.6"}
websockets = "^12.0"
sse-starlette = "^2.0.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

